move logParse func to utils
