+++ cmd/gossamer/commands/utils.go
@@ -4,6 +4,7 @@
 package commands
 
 import (
+	"encoding/json"
 	"fmt"
 	"os"
 	"path/filepath"
@@ -480,4 +481,55 @@ func setViperDefault(config *cfg.Config) {
 			}
 		}
 	}
+}
+
+func parseLogLevel() error {
+	// set default log level
+	moduleToLogLevel := map[string]string{
+		"global":  cfg.DefaultLogLevel,
+		"core":    cfg.DefaultLogLevel,
+		"digest":  cfg.DefaultLogLevel,
+		"sync":    cfg.DefaultLogLevel,
+		"network": cfg.DefaultLogLevel,
+		"rpc":     cfg.DefaultLogLevel,
+		"state":   cfg.DefaultLogLevel,
+		"runtime": cfg.DefaultLogLevel,
+		"babe":    cfg.DefaultLogLevel,
+		"grandpa": cfg.DefaultLogLevel,
+		"wasmer":  cfg.DefaultLogLevel,
+	}
+
+	if len(logLevel) > 0 {
+		logConfigurations := strings.Split(logLevel, ",")
+
+		for _, config := range logConfigurations {
+			parts := strings.SplitN(config, "=", 2)
+			if len(parts) != 2 {
+				return fmt.Errorf("Invalid log configuration: %s", config)
+			}
+
+			module := strings.TrimSpace(parts[0])
+			logLevel := strings.TrimSpace(parts[1])
+
+			if _, ok := moduleToLogLevel[module]; !ok {
+				return fmt.Errorf("Invalid module: %s", module)
+			}
+
+			moduleToLogLevel[module] = logLevel
+		}
+	}
+
+	config.LogLevel = moduleToLogLevel["global"]
+
+	jsonData, err := json.Marshal(moduleToLogLevel)
+	if err != nil {
+		return fmt.Errorf("Error marshalling logs: %s", err)
+	}
+
+	err = json.Unmarshal(jsonData, &config.Log)
+	if err != nil {
+		return fmt.Errorf("Error unmarshalling logs: %s", err)
+	}
+
+	return nil
 }
